JMA exporter uses new collection resources gathering logic
